+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Implement background image sizing with keyword and scale ([#10](https://github.com/marp-team/marpit/pull/10))
 * Support advanced background mode with inline SVG, for multiple images and filters ([#11](https://github.com/marp-team/marpit/pull/11))
 * Upgrade node to the latest LTS version v8.11.1 ([#12](https://github.com/marp-team/marpit/pull/12))
+* Update docs about background images ([#13](https://github.com/marp-team/marpit/pull/13))
 
 ## v0.0.2 - 2018-04-28