+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## v0.0.3 - 2018-05-02
+
 * Implement background image resizing with keyword and scale ([#10](https://github.com/marp-team/marpit/pull/10))
 * Support advanced background mode with inline SVG, for multiple images and filters ([#11](https://github.com/marp-team/marpit/pull/11))
 * Upgrade node to the latest LTS version v8.11.1 ([#12](https://github.com/marp-team/marpit/pull/12))